fix: unable to determine first commit
If first commit on current branch could not be determined then master is taken as fallback.
This assumes that the commits to put into the pull request are commit on master and then a branch is created. It will be inaccurate to chose the correct title and body if there are multiple commits on master, since it is impossible to select the first commit (maybe it would be good to diff against origin/master and select the first commit between orgin/master and master in that case).